+++ src/test/kotlin/com/booleworks/kjobs/control/MaintenanceTest.kt
@@ -41,7 +41,7 @@ class MaintenanceTest : FunSpec({
             addJob(defaultJobType, persistence, { _, _ -> ComputationResult.Success(TestResult(42)) }) {}
         }
 
-        val directCall = { instance: String -> suspend { Maintenance.updateHeartbeat(persistence, instance, null) } }
+        val directCall = { instance: String -> suspend { Maintenance.updateHeartbeat(persistence, instance) } }
         val testingCall = { instance: String -> suspend { testingMode.updateHeartbeat(instance) } }
 
         for (method in listOf(directCall, testingCall)) {